Conflicts:
	app/apps/deskpro_jira/html/Ticket/list.html
	app/apps/deskpro_jira/js/Service/Meta.js

+++ app/src/Application/AgentBundle/Controller/PublishController.php
@@ -233,8 +233,15 @@ class PublishController extends AbstractController
 		}
 
 		$entity = $this->_getCommentEntityName($typename);
+		if (!$entity) {
+			throw $this->createNotFoundException();
+		}
 
 		$comment = $this->em->find($entity, $comment_id);
+		if (!$comment) {
+			throw $this->createNotFoundException();
+		}
+
 		$this->em->remove($comment);
 		$this->em->flush();
